Finish converting navigation components
