+++ src/components/panelContainer/PanelContainer.tsx
@@ -1,7 +1,7 @@
 import {Routes, Route, Navigate} from 'react-router-dom'
 import {TelemetryPanel} from '../telemetry/TelemetryPanel.js'
 // import {HelpPanel} from '../help/HelpPanel.js'
-// import {NavigationPanel} from '../navigation/NavigationPanel.js'
+import {NavigationPanel} from '../navigation/NavigationPanel.js'
 import {ArmDexterityPanel} from '../armDexterity/ArmDexterityPanel.js'
 import './PanelContainer.css'
 import React from 'react'
@@ -12,7 +12,7 @@ function PanelContainer() {
       <Routes>
         <Route path="" element={<Navigate to="/navigation" />} />
 
-        {/* <Route path="/navigation" element={<NavigationPanel />} /> */}
+        <Route path="/navigation" element={<NavigationPanel />} />
 
         <Route path="/arm-dexterity" element={<ArmDexterityPanel />} />
